Move a number of common files out of the unified video tracker folder.
The intent is to share these files, which don't differ (substantially)
between the two video tracker implementations, between the two, so we
don't have needless duplication, code drift, manual merging, etc.
