+++ plugins/videotrackershared/CameraParameters.h
