+++ plugins/videotrackershared/GetOptionalParameter.h
