+++ plugins/videotrackershared/OptionalStream.h
