+++ plugins/videotrackershared/cvToEigen.h
